Fix the chroma oscillator to reset the NCO counter at the beginning of each colour burst.
Without this it would just free-run, leading to the phase of the colour
burst signal being different on each line, leading to all kinds of
colour / dot crawling.

This fixes it to almost completely stop the dot crawling - but it does
leave a tiny amount in very specific cases and a bunch of attribute bleed
which i /think/ is because I'm not QUITE running this at 16/32MHz on my
upduino v3.1 (the PLL can't quite turn 12MHz into it.)

I'm going to need to source/make some 16/32MHz oscillators (and likely
some 27/54MHz ones too for "proper" NTSC/PAL digital signal synth)
for further experiments.