ntsc-test-1: a simple test to (eventually) generate the 3.58MHz oscillator
This is JUST going to be a test to generate the 3.58MHz output that
will be used for colour generation, and ensure that it can generate
a proper, clean in phase signal.

I'm seeing colour artifacts in my board setup that I don't like / trust.
This will hopefully help me diagnose what's up with the 4 bit DAC generation.